\name{LDshrinkCUDA-package}
\alias{LDshrinkCUDA-package}
\alias{LDshrinkCUDA}
\docType{package}
\title{
\packageTitle{LDshrinkCUDA}
}
\description{
\packageDescription{LDshrinkCUDA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LDshrinkCUDA}
\packageIndices{LDshrinkCUDA}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{LDshrinkCUDA}

Maintainer: \packageMaintainer{LDshrinkCUDA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
